Move calibrate interval to a separate module
