+++ lib/sanbase_web/graphql/helpers/utils.ex
@@ -1,154 +1,4 @@
 defmodule SanbaseWeb.Graphql.Helpers.Utils do
-  alias Sanbase.DateTimeUtils
-
-  def calibrate_interval(
-        module,
-        measurement,
-        from,
-        to,
-        interval,
-        min_interval_seconds \\ 300,
-        data_points_count \\ 500
-      )
-
-  def calibrate_interval(
-        module,
-        measurement,
-        from,
-        to,
-        "",
-        min_interval_seconds,
-        data_points_count
-      ) do
-    with {:ok, first_datetime} <- module.first_datetime(measurement) do
-      first_datetime = first_datetime || from
-
-      from =
-        max(
-          DateTime.to_unix(from, :second),
-          DateTime.to_unix(first_datetime, :second)
-        )
-
-      interval =
-        max(
-          div(DateTime.to_unix(to, :second) - from, data_points_count),
-          min_interval_seconds
-        )
-
-      {:ok, DateTime.from_unix!(from), to, "#{interval}s"}
-    end
-  end
-
-  def calibrate_interval(
-        _module,
-        _measurement,
-        from,
-        to,
-        interval,
-        _min_interval,
-        _data_points_count
-      ) do
-    {:ok, from, to, interval}
-  end
-
-  def calibrate_interval(
-        module,
-        metric,
-        slug,
-        from,
-        to,
-        "",
-        min_interval_seconds,
-        data_points_count
-      ) do
-    {:ok, first_datetime} = module.first_datetime(metric, slug)
-
-    first_datetime = first_datetime || from
-
-    from =
-      max(
-        DateTime.to_unix(from, :second),
-        DateTime.to_unix(first_datetime, :second)
-      )
-
-    interval =
-      max(
-        div(DateTime.to_unix(to, :second) - from, data_points_count),
-        min_interval_seconds
-      )
-
-    {:ok, DateTime.from_unix!(from), to, "#{interval}s"}
-  end
-
-  def calibrate_interval(
-        _module,
-        _metric,
-        _slug,
-        from,
-        to,
-        interval,
-        _min_interval,
-        _data_points_count
-      ) do
-    {:ok, from, to, interval}
-  end
-
-  def calibrate_interval_with_ma_interval(
-        module,
-        measurement,
-        from,
-        to,
-        interval,
-        min_interval,
-        ma_base,
-        data_points_count \\ 500
-      ) do
-    {:ok, from, to, interval} =
-      calibrate_interval(module, measurement, from, to, interval, min_interval, data_points_count)
-
-    ma_interval =
-      max(
-        div(
-          DateTimeUtils.str_to_sec(ma_base),
-          DateTimeUtils.str_to_sec(interval)
-        ),
-        2
-      )
-
-    {:ok, from, to, interval, ma_interval}
-  end
-
-  def calibrate_incomplete_data_params(true, _module, _metric, from, to) do
-    {:ok, from, to}
-  end
-
-  def calibrate_incomplete_data_params(false, module, metric, from, to) do
-    case module.has_incomplete_data?(metric) do
-      true -> rewrite_params_incomplete_data(from, to)
-      false -> {:ok, from, to}
-    end
-  end
-
-  defp rewrite_params_incomplete_data(from, to) do
-    start_of_day = Timex.beginning_of_day(Timex.now())
-
-    case DateTime.compare(from, start_of_day) != :lt do
-      true ->
-        {:error,
-         """
-         The time range provided [#{from} - #{to}] is contained in today. The metric
-         requested could have incomplete data as it's calculated since the beginning
-         of the day and not for the last 24 hours. If you still want to see this
-         data you can pass the flag `includeIncompleteData: true` in the
-         `timeseriesData` arguments
-         """}
-
-      false ->
-        to = if DateTime.compare(to, start_of_day) == :gt, do: start_of_day, else: to
-        {:ok, from, to}
-    end
-  end
-
   def error_details(changeset) do
     changeset
     |> Ecto.Changeset.traverse_errors(&format_error/1)
