+++ lib/sanbase_web/graphql/resolvers/project/project_metrics_resolver.ex
@@ -1,7 +1,8 @@
 defmodule SanbaseWeb.Graphql.Resolvers.ProjectMetricsResolver do
   import Sanbase.Utils.ErrorHandling, only: [handle_graphql_error: 3]
   import Absinthe.Resolution.Helpers
-  import SanbaseWeb.Graphql.Helpers.Utils
+
+  import SanbaseWeb.Graphql.Helpers.CalibrateInterval
 
   alias Sanbase.Model.Project
   alias Sanbase.Metric
